chore: move evaluation_details to a new file
