+++ lib/src/evaluation.dart
@@ -1,115 +1,5 @@
-import 'package:bucketeer_flutter_client_sdk/src/value_parser.dart';
 import 'package:flutter/foundation.dart';
-
-@immutable
-class BKTEvaluationDetails<T extends Object> {
-  const BKTEvaluationDetails({
-    required this.featureId,
-    required this.featureVersion,
-    required this.userId,
-    required this.variationId,
-    required this.variationName,
-    required this.variationValue,
-    required this.reason,
-  });
-
-  final String featureId;
-  final int featureVersion;
-  final String userId;
-  final String variationId;
-  final String variationName;
-  final T variationValue;
-  final String reason;
-
-  @override
-  bool operator ==(Object other) {
-    final bool isSameType = other is BKTEvaluationDetails<T>;
-    if (isSameType == false) {
-      return false;
-    }
-    final otherAsBKTEvaluationDetails = other as BKTEvaluationDetails<T>;
-    final bool isRuntimeTypeEqual =
-        runtimeType == otherAsBKTEvaluationDetails.runtimeType;
-    final bool isFeatureIdEqual =
-        featureId == otherAsBKTEvaluationDetails.featureId;
-    final bool isFeatureVersionEqual =
-        featureVersion == otherAsBKTEvaluationDetails.featureVersion;
-    final bool isUserIdEqual = userId == otherAsBKTEvaluationDetails.userId;
-    final bool isVariationIdEqual =
-        variationId == otherAsBKTEvaluationDetails.variationId;
-    final bool isVariationNameEqual =
-        variationName == otherAsBKTEvaluationDetails.variationName;
-    bool isVariationValueEqual =
-        variationValue == otherAsBKTEvaluationDetails.variationValue;
-    final bool isReasonEqual = reason == otherAsBKTEvaluationDetails.reason;
-
-    if (variationValue is Map) {
-      final variationValueMap = variationValue as Map;
-      final otherVariationValueMap = otherAsBKTEvaluationDetails.variationValue as Map;
-      isVariationValueEqual = mapEquals(variationValueMap, otherVariationValueMap);
-    }
-
-    return isSameType &&
-        isRuntimeTypeEqual &&
-        isFeatureIdEqual &&
-        isFeatureVersionEqual &&
-        isUserIdEqual &&
-        isVariationIdEqual &&
-        isVariationNameEqual &&
-        isVariationValueEqual &&
-        isReasonEqual;
-  }
-
-  @override
-  int get hashCode =>
-      featureId.hashCode ^
-      featureVersion.hashCode ^
-      userId.hashCode ^
-      variationId.hashCode ^
-      variationName.hashCode ^
-      variationValue.hashCode ^
-      reason.hashCode;
-
-  @override
-  String toString() {
-    return 'Evaluation{featureId: $featureId, '
-        'featureVersion: $featureVersion, userId: $userId, '
-        'variationId: $variationId, variationName: $variationName, '
-        'variationValue: $variationValue, '
-        'reason: $reason}';
-  }
-
-  static BKTEvaluationDetails<T> fromJson<T extends Object>(
-      Map<String, dynamic> json,
-      {TypeConverter<T>? converter}) {
-    final valueTypeConverter = converter ?? DefaultTypeConverter();
-    return BKTEvaluationDetails<T>(
-      featureId: json['featureId'],
-      featureVersion: json['featureVersion'],
-      userId: json['userId'],
-      variationId: json['variationId'],
-      variationName: json['variationName'],
-      // throw exception if type does not match
-      variationValue: valueTypeConverter.parse(json['variationValue']),
-      reason: json['reason'],
-    );
-  }
-
-  static BKTEvaluationDetails<T> createDefaultValue<T extends Object>(
-      String featureId, String userId, T defaultValue) {
-    return BKTEvaluationDetails<T>(
-      featureId: featureId,
-      featureVersion: 0,
-      userId: userId,
-      variationId: '',
-      variationName: '',
-      variationValue: defaultValue,
-      reason: evaluationDetailsDefaultReason,
-    );
-  }
-
-  static const evaluationDetailsDefaultReason = 'CLIENT';
-}
+import 'evaluation_details.dart';
 
 @Deprecated("use BKTEvaluationDetails<String> instead")
 @immutable
